[PIR #841] Relocate architect-name validator to codev-core (shared single source)
