+++ packages/core/package.json
@@ -32,6 +32,10 @@
       "types": "./dist/agent-names.d.ts",
       "default": "./dist/agent-names.js"
     },
+    "./architect-name": {
+      "types": "./dist/architect-name.d.ts",
+      "default": "./dist/architect-name.js"
+    },
     "./builder-helpers": {
       "types": "./dist/builder-helpers.d.ts",
       "default": "./dist/builder-helpers.js"
